pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Quang Phan <vnphanquang@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
python-dotenv = "^0.20.0"
beautifulsoup4 = "^4.11.1"
redis-om = "^0.0.27"
requests = "^2.28.1"
Flask-Cors = "^3.0.10"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
